Cleaned up generated files from express-generator
