+++ config.json.sample
@@ -1,0 +1,3 @@
+{
+  "slackBotUserApiToken": "SLACKBOTAPITOKEN"
+}
