+++ package.json
@@ -3,19 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "nodemon ./bin/www"
+    "start": "node index.js"
   },
   "dependencies": {
-    "body-parser": "~1.13.2",
-    "cookie-parser": "~1.3.5",
-    "debug": "~2.2.0",
     "express": "~4.13.1",
-    "hbs": "~3.1.0",
-    "morgan": "~1.6.1",
-    "serve-favicon": "~2.3.0",
     "slack-client": "^1.4.1"
-  },
-  "devDependencies": {
-    "nodemon": "^1.8.1"
   }
 }
