--- views/index.hbs
@@ -1,2 +1,0 @@
-<h1>{{title}}</h1>
-<p>Welcome to {{title}}!</p>
